Update gemspec and bump version to 0.7.0 to reflect focus on a single datastore.
Overtime, the resemblance of the implementation to quasor's
original has drifted significantly. All that's used from this implementation
is the redis + diskstore approach, and the others meant loading a
bunch of gems and code that isn't necessary.

This is the slimmer version. It will retain wireline compatibility
with statsd, but not datastore - it won't go to graphite.